TIP-236: Merge repositories
