+++ src/Pim/Bundle/UserBundle/Doctrine/ORM/Repository/GroupRepository.php
@@ -2,9 +2,11 @@
 
 namespace Pim\Bundle\UserBundle\Doctrine\ORM\Repository;
 
-use Akeneo\Component\StorageUtils\Repository\IdentifiableObjectRepositoryInterface;
+use Doctrine\ORM\EntityRepository;
+use Doctrine\ORM\QueryBuilder;
 use Pim\Bundle\UserBundle\Entity\User;
-use Pim\Bundle\UserBundle\Repository\GroupRepository as BaseGroupRepository;
+use Pim\Component\User\Model\GroupInterface;
+use Pim\Component\User\Repository\GroupRepositoryInterface;
 
 /**
  * User group repository
@@ -13,15 +15,32 @@ use Pim\Bundle\UserBundle\Repository\GroupRepository as BaseGroupRepository;
  * @copyright 2014 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class GroupRepository extends BaseGroupRepository implements
-    IdentifiableObjectRepositoryInterface
+class GroupRepository extends EntityRepository implements GroupRepositoryInterface
 {
+    /**
+     * {@inheritdoc}
+     */
+    public function getIdentifierProperties()
+    {
+        return ['name'];
+    }
+
     /**
      * {@inheritdoc}
      */
     public function findOneByIdentifier($code)
     {
         return $this->findOneBy(['name' => $code]);
+    }
+
+    /**
+     * Get the default user group
+     *
+     * @return null|object
+     */
+    public function getDefaultUserGroup()
+    {
+        return $this->findOneByIdentifier(User::GROUP_DEFAULT);
     }
 
     /**
@@ -47,20 +66,19 @@ class GroupRepository extends BaseGroupRepository implements
     }
 
     /**
-     * Get the default user group
+     * Get user query builder
      *
-     * @return null|object
+     * @param  GroupInterface $group
+     *
+     * @return QueryBuilder
      */
-    public function getDefaultUserGroup()
+    public function getUserQueryBuilder(GroupInterface $group)
     {
-        return $this->findOneByIdentifier(User::GROUP_DEFAULT);
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function getIdentifierProperties()
-    {
-        return ['name'];
+        return $this->_em->createQueryBuilder()
+            ->select('u')
+            ->from('PimUserBundle:User', 'u')
+            ->join('u.groups', 'groups')
+            ->where('groups = :group')
+            ->setParameter('group', $group);
     }
 }
