--- src/Pim/Bundle/UserBundle/Repository/RoleRepository.php
@@ -1,25 +1,0 @@
-<?php
-namespace Pim\Bundle\UserBundle\Repository;
-
-use Doctrine\ORM\EntityRepository;
-use Doctrine\ORM\QueryBuilder;
-use Pim\Component\User\Model\Role;
-
-class RoleRepository extends EntityRepository
-{
-    /**
-     * Get user query builder
-     *
-     * @param  Role         $role
-     * @return QueryBuilder
-     */
-    public function getUserQueryBuilder(Role $role)
-    {
-        return $this->_em->createQueryBuilder()
-            ->select('u')
-            ->from('PimUserBundle:User', 'u')
-            ->join('u.roles', 'role')
-            ->where('role = :role')
-            ->setParameter('role', $role);
-    }
-}
